Add erasure coding
Unconditionally pads the data to a whole number of QR codes. Adds a length parameter at the start of the data. Un-pads on restore using the length.

Adds --erasure-coding, --no-erasure-coding. Erasure coding is on by default. It is always set to 30% acceptable loss, and can't be configured.

Add --shuffle, --no-shuffle. The default is to turn shuffle on only when erasure coding is enabled.