change note font size to a monospace font
The DejaVu font is really ambiguous. For example, the O and 0 look
very similar, which is a problem when you're transcribing technical
documents. This is especially a problem when including a textual copy
of the backups in the additional notes (#48).

We use DejaVu Sans Mono to avoid introducing a new dependency but
maybe a better option would be to make this a user-facing option. I
looked into that as well, but dishartened by the custom options
parsed (have you considered using argparse? :).