reduce severity of verification warning and clarify warning
I was surprised to see this message when I tried qr-backup. I thought
the entire process failed because there was this message (and only
it):

    CRITICAL: Skipping digital restore verification, because 'convert' is not available. Debian/Ubuntu forbid PDF conversion using imagemagick. More information at: https://github.com/za3k/qr-backup/tree/master/docs

Normally, critical errors are irrevocable that basically stop the
entire process. And indeed, in other places where we use logging.fatal
in the program, we immediately exit.

So let's turn this into a WARNING instead.

I've also considered explicitely mentioning that the file was
correctly converted, but this seems to me like a message that should
come before, if --verbose is enabled. Unfortunately, --verbose looks
more like a --debug to me and outputs way too much information. :) So
I'm not sure how to fix that particular limitation (that we don't tell
the user the file was correctly generated).

We also indicate more clearly which part of the documentation we're
refering to here.

Finally, also remove the explicit reference to Debian/Ubuntu, as that
seriously limits the scope of the warning. In fact, any Debian
derivative (not just Ubuntu!) is going to have this problem, and I
believe other distributions (e.g. Fedora, Red Hat) have implemented
similar workarounds.