Use TIF instead of PDF.
1. Previously, we used PDFs to store the data, which `zbarimg` can't
interpret reliably.  Now, we store the data in multi-page TIF files
which are significantly smaller files that `zbarimg` can reliably
read.  The critical change is on line 1281, where I specified
`format="tiff"` and a compression scheme.

2. Additionally, I changed the default output.  Now, instead of printing
to stdout by default, we save to a file.  This saves the user from
needing to pick the file extension and unintentionally write to
unreliable PDFs.  Now, you must use "-" to manually specify outputting
to stdout.  The critical change is on line 1110.

With these two changes, I'm able to write and restore 750KB
backups (bz2 compressed to 200KB) in 40 pages.  The rest of the
changes are related to renaming variables to be more generic and
fixing the relevant help-text.

I recommend testing these modes with this backup command:

    rm qr-backup.log
    time python3 ./qr-backup -v --instructions cover ~/test.tar.bz2 \
    2>&1 | tee qr-backup.log

And this restore command:

    rm qr-restore.log
    time python3 ./qr-backup -v --restore ~/test.tar.bz2.qr.tif -o \
    ~/test.tar.bz2 2>&1 | tee qr-restore.log