Print qr-backup failures in tests
The tests (in master) were failing with the alarming (and opaque)
message "incorrect-restore". However the problem was simply that the author
is a dum-dum who didn't install qr-backup's command line dependencies. To
prevent this problem from recurring, the following changes were added to
the test infrastructure:

- If qr-backup is failing because of missing CLI dependencies, the tests now
report the specific missing packages.

- If qr-backup fails with a non-zero return code, the error is now reported
differently than generic backup/regression failures, and the error (stderr)
is printed as part of the test output.

Note that there is still a regression failure in master, but the author is
still dum-dum so its actual existence has not yet been confirmed.