Fix failing reproducibility tests in master.
Overall the backups are working fine, so all this commit does it update blessed values and add some additional tests.

First, this should never have gotten to master for more than a day. There is an ongoing PR https://github.com/za3k/qr-backup/pull/64 to add continuous integration, so that if tests fail we are immediately alerted. (and for PRs, we know if tests pass before merging)

On to the actual reproducibility issues.

The first is that the explanatory text changed from
    "To restore this file from the command line in Linux" to
    "To restore this file from the command line"

The second is that python's gzip module had a reproduciblity regression between 3.10 and 3.11 where one byte differed in output (and generally does not guarantee this will remain identical). See the python bug at https://github.com/python/cpython/issues/112346 for details.