Add --fast option to tests
The new option

(1) Runs tests in parallel
(2) Omits tests which are intended for developers to spot reproducibility issues, rather than for end-users to verify a working install. This means --fast is also more reliable.

The opposite of --fast is --full.

Note that we're using a custom test runner, which is not ideal. It's because we're using generated tests, and the generated tests share a lot of redundant setup. But it's still not ideal.

Closes issue #67.