Disable self-check by default.
I'm going to disable self-test as a whole for the time being. It was a good idea in theory, but it's a bad user experience in the current state to get constant false positives.

It was a mistake that I left it on even a tenth this long. I think I had the vague idea that it would get me/others to fix zbarimg's reliability, but that logic was clearly wrong (and was not user-centric enough)